+++ tasmota/tasmota_xdrv_driver/xdrv_11_knx.ino
@@ -644,7 +644,7 @@ void KNX_CB_Action(message_t const &msg, void *arg)
           knx.answer_4byte_float(msg.received_on, last_hum);
         }
       }
-#if defined(USE_ENERGY_SENSOR)      
+#if defined(USE_ENERGY_SENSOR)
       else if (chan->type == KNX_ENERGY_VOLTAGE) // Reply KNX_ENERGY_VOLTAGE
       {
         knx.answer_4byte_float(msg.received_on, Energy.voltage[0]);
@@ -1298,7 +1298,7 @@ void CmndKnxCb(void)
  * Interface
 \*********************************************************************************************/
 
-bool Xdrv11(uint8_t function)
+bool Xdrv11(uint32_t function)
 {
   bool result = false;
     switch (function) {
