+++ tasmota/tasmota_xdrv_driver/xdrv_50_filesystem.ino
@@ -825,7 +825,7 @@ void UfsListDir(char *path, uint8_t depth) {
           editpath[0]=0;
   #endif // GUI_TRASH_FILE
           ext_snprintf_P(npath, sizeof(npath), UFS_FORM_SDC_HREF, ppe, epe);
-          WSContentSend_P(UFS_FORM_SDC_DIRb, hiddable ? UFS_FORM_SDC_DIR_HIDDABLE : UFS_FORM_SDC_DIR_NORMAL, npath, epe, 
+          WSContentSend_P(UFS_FORM_SDC_DIRb, hiddable ? UFS_FORM_SDC_DIR_HIDDABLE : UFS_FORM_SDC_DIR_NORMAL, npath, epe,
                           HtmlEscape(name).c_str(), tstr.c_str(), entry.size(), delpath, editpath);
         }
         entry.close();
@@ -1109,7 +1109,7 @@ void UfsEditorUpload(void) {
  * Interface
 \*********************************************************************************************/
 
-bool Xdrv50(uint8_t function) {
+bool Xdrv50(uint32_t function) {
   bool result = false;
 
   switch (function) {
