+++ tasmota/tasmota_xsns_sensor/xsns_02_analog.ino
@@ -820,7 +820,7 @@ void CmndAdcParam(void) {
  * Interface
 \*********************************************************************************************/
 
-bool Xsns02(uint8_t function) {
+bool Xsns02(uint32_t function) {
   bool result = false;
 
   switch (function) {
